requirements.txt
python-telegram-bot>=13.7
celery>=5.2.0
SQLAlchemy>=1.4.26
psycopg2
redis